requirements.txt
textstat
japanize_matplotlib
xfeat
category_encoders
texthero
transformers
torchinfo